refactor: move class CalendarExport to namespace Backend
